+++ binding.gyp
@@ -4,16 +4,16 @@
       'target_name': 'profiler',
       'win_delay_load_hook': 'false',
       'sources': [
+        'src/cpu_profiler/cpu_profiler.cc',
+        'src/cpu_profiler/cpu_profile.cc',
+        'src/cpu_profiler/cpu_profile_node.cc',
+        'src/heap_profiler/sampling_heap_profiler.cc',
+        'src/heapsnapshot/heap_profiler.cc',
+        'src/heapsnapshot/heap_snapshot.cc',
+        'src/heapsnapshot/heap_output_stream.cc',
+        'src/heapsnapshot/heap_graph_node.cc',
+        'src/heapsnapshot/heap_graph_edge.cc',
         'src/profiler.cc',
-        'src/cpu_profiler.cc',
-        'src/cpu_profile.cc',
-        'src/cpu_profile_node.cc',
-        'src/heap_profiler.cc',
-        'src/heap_snapshot.cc',
-        'src/heap_output_stream.cc',
-        'src/heap_graph_node.cc',
-        'src/heap_graph_edge.cc',
-        'src/sampling_heap_profile.cc',
         'src/environment_data.cc'
       ],
       'include_dirs' : [
