+++ lib/worker_threads.js
@@ -8,7 +8,7 @@ try {
   const workerThreads = require('worker_threads');
   isMainThread = workerThreads.isMainThread;
   threadId = workerThreads.threadId;
-  canIUseWorkerThreads = true;
+  // canIUseWorkerThreads = true;
 } catch (err) {
   err;
 }
