+++ src/heapsnapshot/heap_output_stream.cc
