+++ src/heapsnapshot/heap_output_stream.h
